python-app/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
python-dotenv
sentence-transformers
torch
transformers
scikit-learn
scipy
